+++ src/app/code/community/Dhl/Versenden/Model/Adminhtml/System/Config/Source/Procedure.php
@@ -36,7 +36,7 @@ class Dhl_Versenden_Model_Adminhtml_System_Config_Source_Procedure
         return array(
             // DHL Paket procedures
             Product::PROCEDURE_PAKET_NATIONAL => $helper->__('DHL Paket: V01PAK'),
-            Product::PROCEDURE_WARENPOST_NATIONAL => $helper->__('DHL Warenpost National: V62WP'),
+            Product::PROCEDURE_KLEINPAKET => $helper->__('DHL Kleinpaket: V62KP'),
             Product::PROCEDURE_WELTPAKET => $helper->__('DHL Paket International: V53WPAK'),
             Product::PROCEDURE_WARENPOST_INTERNATIONAL => $helper->__('DHL Warenpost International: V66WPI'),
             Product::PROCEDURE_RETURNSHIPMENT_NATIONAL => $helper->__('Retoure DHL Paket: V01PAK'),