+++ CHANGELOG.md
@@ -4,6 +4,19 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## 1.14.0 - 2025-01-14
+
+Deutsche Post Warenpost becomes DHL Kleinpaket
+
+### Added
+
+- support for DHL Kleinpaket
+
+### Removed
+
+- support for Deutsche Post Warenpost
+
+
 ## 1.13.1 - 2024-09-25
 
 ### Changed